Change int to uint8_t in various for loops.

+++ esphome/components/lc709203f/lc709203f.cpp
@@ -174,7 +174,7 @@ uint8_t Lc709203f::get_register_(uint8_t register_to_read, uint16_t *register_va
   read_buffer[1] = register_to_read;
   read_buffer[2] = ((this->address_) << 1) | 0x01;
 
-  for (int i = 0; i <= LC709203F_I2C_RETRY_COUNT; i++) {
+  for (uint8_t i = 0; i <= LC709203F_I2C_RETRY_COUNT; i++) {
     // Note: the read_register() function does not send a stop between the write and
     //  the read portions of the I2C transation when you set the last variable to 'false'
     //  as we do below. Some of the other I2C read functions such as the generic read()
@@ -223,7 +223,7 @@ uint8_t Lc709203f::set_register_(uint8_t register_to_set, uint16_t value_to_set)
   write_buffer[3] = (value_to_set >> 8) & 0xFF;  // High byte
   write_buffer[4] = this->crc8_(write_buffer, 4);
 
-  for (int i = 0; i <= LC709203F_I2C_RETRY_COUNT; i++) {
+  for (uint8_t i = 0; i <= LC709203F_I2C_RETRY_COUNT; i++) {
     // Note: we don't write the first byte of the write buffer to the device.
     //  This is done automatically by the write() function.
     return_code = this->write(&write_buffer[1], 4, true);
@@ -245,10 +245,10 @@ uint8_t Lc709203f::crc8_(uint8_t *byte_buffer, uint8_t length_of_crc) {
   uint8_t crc = 0x00;
   const uint8_t polynomial(0x07);
 
-  for (int j = length_of_crc; j; --j) {
+  for (uint8_t j = length_of_crc; j; --j) {
     crc ^= *byte_buffer++;
 
-    for (int i = 8; i; --i) {
+    for (uint8_t i = 8; i; --i) {
       crc = (crc & 0x80) ? (crc << 1) ^ polynomial : (crc << 1);
     }
   }
@@ -265,7 +265,7 @@ void Lc709203f::set_pack_size(uint16_t pack_size) {
 
   // The size is used to calculate the 'Adjustment Pack Application' number.
   // Here we assume a type 01 or type 03 battery and do a linear curve fit to find the APA.
-  for (int i = 0; i < sizeof(PACK_SIZE_ARRAY) / sizeof(PACK_SIZE_ARRAY[0]); i++) {
+  for (uint8_t i = 0; i < sizeof(PACK_SIZE_ARRAY) / sizeof(PACK_SIZE_ARRAY[0]); i++) {
     if (PACK_SIZE_ARRAY[i] == pack_size) {
       // If the pack size is exactly one of the values in the array.
       this->apa_ = APA_ARRAY[i];
